Create a page-specific store for handling editing a triple in the entity page
This means that editing is scoped to the entity page for now, and leaving will remove all the triples from the local store. This also means that editing and publishing can only happen one page at a time. This is probably fine until we add a mechanism for viewing all local changes.