Correctly merge entities that have changed locally
This is a separate workflow from entities that are created locally. Since changed entities may not have had _all_ of their triples changed, we could be missing data required for rendering the table when we do the merge.

Now we fetch entity triples if we do not have the data for it already.