feat: add additional toast step for signing tx during publish
Previously there was no step in the publish toast telling the user to sign the outgoing publish transaction in their wallet. Additionally, sometimes the notification for signing the notification would never appear in the wallet, leaving users in a purgatory state where they couldn't sign the transaction and couldn't retry without refreshing the page and losing changes.

This feature handles now enables users to retry the transaction prompt. In the future we will investigate the root cause for missing tx notification and also store user changes in persistent storage.