fix: stale name cell (#220)
For some reason the table value for the name field is stale when changing the selectedType in edit mode. When debugging it looks like the cell has the correct data, but the value of the name is the value of the last cell in the previous selectedType. For now we can use a key to force the cell to re-mount when the selectedType and name changes.