Use proposedVersions for the history list instead of finalized Versions
This is because the finalized Version has _all_ changes that have been made to an entity from genesis to the current state. We want the _diff_ between the last state and the next version, which is what Proposed Version provides. Once we have governance we want to make sure we only query for _approved_ ProposedVersions and not _all_ ProposedVersions.