feat: query profile information and map to version createdBy
Right now there is no wallet -> profile mapping in the subgraph until we added permissionless spaces and profiles. We can create an ad-hoc profile system using GeoEntities and associated them with a Wallet entity.

This PR sets up infrastructure for querying a Profile GeoEntity based on the wallet address in a version's createdBy field. We then add the profile information to the version to populate the name and avatar.