Fetch as many versions as possible
Previously we were restricting the number of versions at the query level. This would result in incorrect contributor count information since the number of contributors would only be counted from the queries that we made. We now default to the default graph-node query count and take the first 10 versions on the client for the history UI.