feat: merge profile with versions on the server when fetching proposed versions (#242)
* feat: query profile information and map to version createdBy

Right now there is no wallet -> profile mapping in the subgraph until we added permissionless spaces and profiles. We can create an ad-hoc profile system using GeoEntities and associated them with a Wallet entity.

This PR sets up infrastructure for querying a Profile GeoEntity based on the wallet address in a version's createdBy field. We then add the profile information to the version to populate the name and avatar.

* Update avatar group in rest of app

* Fix bug in actions mapping

Sometimes attribute can return null. This is probably a side-effect from old data. Needs more investigation.

* Use system ids for hardcoded attribute indexing in subgraph

* Fix build error

* Fetch as many versions as possible

Previously we were restricting the number of versions at the query level. This would result in incorrect contributor count information since the number of contributors would only be counted from the queries that we made. We now default to the default graph-node query count and take the first 10 versions on the client for the history UI.

* Update comment on recent version parsing

* Migrate Avatar design system component to handle profile Avatars