hide navbar actions until we are on the client
Hide navbar actions until we are on the client. This is because our account state only exists on the client due to the nature of wallets. By having different client and server states on first render we trigger hydration errors.

One possible solution is to track login state as a cookie, but for now we don't track any login state on the server.

We encapsulate the search in the ClientOnly even though its not dependent on account state so we don't get any layout shift when the navbar actions appear.