feat: profiles navbar (#413)
* simplify route segment reading using new next 13 hooks

* pass correct url segment to navbar actions

* Use different button when disconnected

* Avatar menu

* naive local profile hook

revert use access control hack

* let parent control button text styling

let parent control button text styling

* add toggle animations

* toggle state was reversed

* move to animated toggle pill component

* use smiley default avatars

* add additional items to avatar dropdown

* add disabled state to dropdown items

* faster flow-bar animation to align with other animations in app

* add shake animation if user does not have edit access

* clean up edit toggling

* hide navbar actions until we are on the client

Hide navbar actions until we are on the client. This is because our account state only exists on the client due to the nature of wallets. By having different client and server states on first render we trigger hydration errors.

One possible solution is to track login state as a cookie, but for now we don't track any login state on the server.

We encapsulate the search in the ClientOnly even though its not dependent on account state so we don't get any layout shift when the navbar actions appear.

* naive implementation of showing the tooltip

* only handle shaking and tooltip logic in the context of a space

* cleanup connected/disconnected functionality in geo connect button