better animation
There's some jank between the error and non-error states. If we have popLayout the layout animation works, but there's some residual content showing before going away. If we use wait then there's jank in the layout animation. We might have to consolidate the content of error and non-error so we don't change the actual motion div being rendered.