feat: animate status bar (#471)
* animate status bar

* better bounce

* visually center error text vertically

* better animation

There's some jank between the error and non-error states. If we have popLayout the layout animation works, but there's some residual content showing before going away. If we use wait then there's jank in the layout animation. We might have to consolidate the content of error and non-error so we don't change the actual motion div being rendered.

* i fixed it

* fix final state

* Revert "i fixed it"

This reverts commit 182e94cdecfad265641bb51879e1dabbc38e1029.

* Revert "Revert "i fixed it""

This reverts commit 1ae0aca4d43ad2156b4f58f3800ba68666a0443d.

* slide up should not be a dialog component