fix: clean up attribute relation value type implementation
This PR adds back reverted logic for squashing actions before review time. Additionally, it adds code to handle cases where we are still fetching the space for the attribute id. Lastly it works around an edge-case in `fromActions` where merging an already merged set of triples will result in duplicates.