fix: dedupe triples at the fromActions layer
We might be merging actions into a set of triples that have already been merged. In this case we need to replace the existing triple instead of adding a new one. Otherwise we will get duplicate triples since we've added the `createTriple` action multiple times.

For now we dedupe the entire triples array at the end. If this becomes a performance problem we can instead dedupe at the `createTriple` section of the switch statement in `fromActions`.