feat: migration system (#488)
* feat: initial migration hub

The migration hub listens for events that will trigger a specific type of migration, e.g., a value type change or a deleted entity.

* feat: remove triples that reference deleted entity one-level deep

For now we only do one-level-deep deletes for triples referencing a deleted entity. Eventually we might do cascading deletes which will require traversing the Graph for all nodes that are children of the entity.

* fix: use forEach instead of map

* feat: paginate queries for triple entries greater than 1000

graph-node has a max of 1000 entries returned from the API. If there are more than 1000 triple references we need to paginate through the API until the end. Unfortunately there's no way to know the total number of entries.

* fix: remove test migration page

* fix: update type usage with ValueTypeId