feat: view members of space (#494)
* fix: move space metadataheader to spaces-page partials

* feat: space editors chip

* feat: move space page components up a level

* feat: force caching and prefetching in prod (#493)

* feat: force caching and prefetching in prod

Set different page configurations based on whether we are in dev or prod.

Next will not cache or prefetch at the fetch level when using functions that opt-out of dynamic page caching, such as `cookies()` and `headers()`. We only  use cookies in pages in dev mode, so in prod we can force cache and prefetch  at the fetch level.

* feat: add nextjs docs link

* fetch members

* fix: removed editable state from editor

* feat: move entity page components up a level

* feat: tabs for space overview and governance

* feat: shared layout for space overview and space governance

* fix: space overview page using wrong components

* fix(hack): only delay layout animation in templates

* fix: use regular react component layout instead of nextjs layout

We don't want this layout to nest/compose with child layouts.

* fix: remove animation fixes

This isn't working as intended right now. Will fix on a future polish pass

* fix: remove unused mock data