feat: add noUncheckedIndexAccess to tsconfig
This is a safer way to deal with reading potentially undefined items from arrays and objects. One downside is that TypeScript is not able to do type narrowing when checking object access directly, so you can have situations with more verbose, defensive code than you would otherwise. It's probably a good tradeoff for the correct types.