feat(wip): refactor mapping in prep for queue implementation
Previously populateX functions were doing many things at once
- Writing to the cache the current cursor
- Collecting database entities to write to postgres
- Writing the entities to postgres (in a slow, blocking way)

Now we still collect and write (in a non-blocking, but not super fast way), but it's set up to where we can return the data we are intending to write to stash in a queue for future processing.

We need a queue so we can ensure that operations are completed in the correct order. Right now we have a weird promise then-able queue which isn't optimized.