fix: fix indexing issues with new substream indexer (#590)
* feat: rewrite entry point

* feat: add pg types

* feat: use runStream from substreams package

* feat: remove comment

* fix: cleanup

* fix: peer deps for substream

* fix: datauri in full entry should be of type ZodAction

* fix: correctly bootstrap root entities

* fix: populate proposals

* feat: revert initial genesis block

* feat: read start block from env if it exists, fallback to constant

* feat: misc cleanup

* feat: more misc cleanup

* feat: update tsconfig with optimal settings

* fix: don't overwrite entity column values with null

There may be instances where some of the entity column values aren't changed as part of a set of actions. We don't want to overwrite those entity columns with null values.

* feat: improve mapping/upsert performance

* feat: TODO

* feat: standardize npm scripts for running substream

* feat(wip): refactor mapping in prep for queue implementation

Previously populateX functions were doing many things at once
- Writing to the cache the current cursor
- Collecting database entities to write to postgres
- Writing the entities to postgres (in a slow, blocking way)

Now we still collect and write (in a non-blocking, but not super fast way), but it's set up to where we can return the data we are intending to write to stash in a queue for future processing.

We need a queue so we can ensure that operations are completed in the correct order. Right now we have a weird promise then-able queue which isn't optimized.

* fix: handle string_value for all value types

* feat: handle entity name and description in triple processing

We upsert entity names and descriptions during triple processing since triples are aware of the full, ordered state of actions. This ensures we correctly delete names/descriptions or correctly create them based on the order of the actions.

* fix: actually run the queries for entity name/description

* feat: add eslint + prettier

* feat: add eslint + prettier to api package

* fix: lockfile

* fix: lockfile

* feat: add deployment configuration for substream and API

* fix: this shouldn't work

* fix: try moving eslint import to dev deps

* feat: merge lockfile from main

* fix: update eslint env

* fix: apparently need to use es2021 for some reason

* fix: upgrade typescript eslint parser for newer typescript versions

* fix: update eslint plugin

* fix: eslint errors after upgrades

* fix: allow explicit any