feat: disable dependent data block queries until their dependent data exists
Many of our queries in the data blocks are dependent on other queries, so we get this cascading waterfalls of queries that are all running while waiting for previous queries to finish. This is why we get the layout shift from loading states -> empty states -> loading states before finally resolving.

We now use the enabled flag in useQuery to not run dependent queries until the required data actually exists so it's a lot smoother. Also reducers the number of queries we're making dramatically.