feat: lift error handling to handleMessage caller
Most of the errors in the substream fall into a couple categories and can mostly be handled the same way
1. A space doesn't exist for a given plugin/dao address
2. Inserting into the db fails for some reason

We now bubble these errors to the caller so we can handle the same category of errors in a single place instead of ad-hoc throughout the codebase.

Next we'll need to handle gracefully failing in places where that's okay.