feat: add dao address to all "completed" events
We now emit the dao address for all completed events. e.g., adding a member, removing a subspace, published edits. This makes is so we can do a lookup on the dao itself to make sure the plugin is valid for the action instead of having to look up the dao based on the plugin.