fix: keep table async data from previous state while computing new state (#910)
Previously changing any state that required async requests would send the table into a loading state when making the changes. This was janky. Now we keep previous state while computing the new state. This still isn't perfect as there might be a delay while we wait for your local state to show up in the UI, but it's better than what we had previously.