feat: write entity to relations
Previously we only stored the version for type/from/to. We now also store the entity for each field. This is to support use-cases where you always want the latest version of type/from/to _and_ use cases where you need the version at the time the version was created.