feat(substream): store entities on relation data model (#1032)
* fix: remove relations when deleting entity

* feat: write entity to relations

Previously we only stored the version for type/from/to. We now also store the entity for each field. This is to support use-cases where you always want the latest version of type/from/to _and_ use cases where you need the version at the time the version was created.

* feat: update queries for live relation use-cases

* fix: unused comment

* fix: unused log

* fix: correctly write types to version

* fix: use historical version for previous versions in diff

* feat: update log structure

* feat: new indexes

* feat: use hypergraph API